Trait backup_restore::adapter::Adapter

pub trait Adapter {
#[must_use]
    fn write_new_file<'life0, 'async_trait>(
        &'life0 self, 
        content: impl StreamExt<Item = Vec<u8>> + Send + 'async_trait
    ) -> Pin<Box<dyn Future<Output = Result<FileHandle>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn read_file_content(file_handle: &FileHandle) -> BoxStream<Result<Vec<u8>>>;
}

Adapter defines the interfaces of the storage backend we use for backup and restore.

Required methods

#[must_use] fn write_new_file<'life0, 'async_trait>(
&'life0 self,
content: impl StreamExt<Item = Vec<u8>> + Send + 'async_trait
) -> Pin<Box<dyn Future<Output = Result<FileHandle>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Writes the content of the file to the storage backend and returns a handle to the file once finished.

fn read_file_content(file_handle: &FileHandle) -> BoxStream<Result<Vec<u8>>>

Returns the content of the file in a stream.

Loading content...